requirements.txt
absl-py==0.11.0
asgiref==3.3.1
astunparse==1.6.3
attrs==20.3.0
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
dill==0.3.3
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
h5py==2.10.0
idna==2.10
importlib-resources==3.3.0
joblib==0.17.0
Keras-Preprocessing==1.1.2
Markdown==3.3.3
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pdfminer==20191125
Pillow==8.0.1
promise==2.3
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.8
pytz==2020.4
regex==2020.10.28
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sqlparse==0.4.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
tensorflow-metadata==0.25.0
termcolor==1.1.0
tfds-nightly==4.1.0.dev202012020108
tqdm==4.51.0
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
